Move top-level files and setup codecov.yml
