+++ .storybook/config.js
@@ -2,7 +2,7 @@ import { configure, addDecorator } from '@storybook/react';
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import * as React from 'react';
-import handleAction from '../src/reducer';
+import handleAction from '../src/utils/reducer';
 
 function loadStories() {
   require('../src/components/stories');
@@ -11,7 +11,7 @@ function loadStories() {
   req.keys().forEach(req);
 }
 
-addDecorator(story => (
+addDecorator((story) => (
   <Provider store={createStore(handleAction)}>
     {story()}
   </Provider>
