+++ src/components/stats/ModulePanelContainers.js
@@ -2,14 +2,14 @@
  * @flow
  */
 
-import type { Dispatch } from '../../reducer';
+import type { Dispatch } from '../../utils/reducer';
 import type { DispatchProps } from './ModulePanels';
 import type {ExtendedModule} from '../../types/Stats';
 
 import {connect} from 'react-redux';
 import {
   FocusElement,
-} from '../../actions';
+} from '../../utils/actions';
 import {
   RequiredByPanel,
   RequirementsPanel,
@@ -34,5 +34,3 @@ export const RequirementsPanelContainer = connect(
   null,
   mapDispatchToProps,
 )(RequirementsPanel);
-
-
