+++ src/utils/__tests__/flatten.test.js
